refactor: 디자인 변경 및 패키지 분리
