+++ docs/README.md
@@ -59,7 +59,8 @@
   - 두 숫자에 대해 스트라이크와 볼을 계산한다.
     - 같은 자리에서 숫자를 맞추면 스트라이크
     - 다른 자리지만 숫자가 들어있다면 볼
-- 컨트롤러
+
+- 진행
   - "숫자 야구 게임을 시작합니다."라는 안내를 출력한다.
   - "숫자를 입력해주세요 : "라는 안내를 출력한다.
   - 사용자 입력에 대해 예외를 처리한다.
