+++ src/main/java/baseball/Application.java
@@ -1,6 +1,6 @@
 package baseball;
 
-import static Controller.Execution.*;
+import static baseball.Execution.*;
 
 public class Application {
 	public static void main(String[] args) {
