+++ src/main/java/baseball/utils/Converter.java
@@ -1,4 +1,4 @@
-package Controller;
+package baseball.utils;
 
 import java.util.List;
 import java.util.stream.Collectors;
