+++ src/main/java/baseball/view/Printer.java
@@ -1,16 +1,16 @@
-package Controller;
+package baseball.view;
 
-import static User.Outputs.*;
+import baseball.constant.ViewConstants;
 
 public class Printer {
 	private static final String SPACE = " ";
 
 	public static void printGameStartMessage() {
-		System.out.println(GAME_START_MESSAGE.get());
+		System.out.println(ViewConstants.GAME_START_MESSAGE.get());
 	}
 
 	public static void printAskingInputMessage() {
-		System.out.print(ASKING_INPUT_MESSAGE.get());
+		System.out.print(ViewConstants.ASKING_INPUT_MESSAGE.get());
 	}
 
 	public static void printResult(int ballsCount, int strikesCount) {
@@ -31,24 +31,24 @@ public class Printer {
 	}
 
 	public static void printStrikesCountMessage(int strikesCount) {
-		System.out.printf(STRIKES_COUNT_MESSAGE.get(), strikesCount);
+		System.out.printf(ViewConstants.STRIKES_COUNT_MESSAGE.get(), strikesCount);
 	}
 
 	public static void printBallsCountMessage(int ballsCount) {
-		System.out.printf(BALLS_COUNT_MESSAGE.get(), ballsCount);
+		System.out.printf(ViewConstants.BALLS_COUNT_MESSAGE.get(), ballsCount);
 	}
 
 	public static void printNothingMessage() {
-		System.out.println(NOTHING_MESSAGE.get());
+		System.out.println(ViewConstants.NOTHING_MESSAGE.get());
 	}
 
 	public static void printGettingRightAnswerMessage(int lengthOfNumber) {
-		System.out.printf(GETTING_RIGHT_ANSWER_MESSAGE.get(), lengthOfNumber);
+		System.out.printf(ViewConstants.GETTING_RIGHT_ANSWER_MESSAGE.get(), lengthOfNumber);
 		System.out.println();
 	}
 
 	public static void printRestartOrNotMessage(String restartGame, String endGame) {
-		System.out.printf(RESTART_OR_NOT_MESSAGE.get(), restartGame, endGame);
+		System.out.printf(ViewConstants.RESTART_OR_NOT_MESSAGE.get(), restartGame, endGame);
 		System.out.println();
 	}
 }
